FInished publications preview section
