+++ packages/blog-starter-kit/themes/modern/components/publicationsPosts/publication-posts.module.scss
@@ -1,0 +1,6 @@
+.grid {
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr;
+    place-items: center;
+    row-gap: 1em;
+}
