+++ packages/blog-starter-kit/themes/modern/pages/tag/[slug].tsx
@@ -15,7 +15,7 @@ import {
 } from '../../generated/graphql';
 import ExternalLinkSVG from '../../components/icons/svgs/ExternalLinkSVG';
 import { createHeaders, createSSRExchange, getUrqlClientConfig } from '../../lib/api/client';
-import PublicationPosts from '../../components/publication-posts';
+import PublicationPosts from '../../components/publicationsPosts/publication-posts';
 import PublicationFooter from '../../components/publication-footer';
 
 const INITIAL_LIMIT = 6;
